feat(function): LoopTrap split → LoopVar + LoopLet (one mode each, bridge prose)
Decompose §2's bundled LoopTrap (mode toggle + stepper) into two single-mode
widgets. LoopVar shows five callbacks tethered to one shared cell; LoopLet
shows five callbacks tethered to five per-iteration cells. The var-vs-let
contrast moves into the connecting prose between them. ModeToggle deleted
(no remaining callers in this commit). §4 cross-reference rewritten from
"scroll back up to LoopTrap" to "scroll back up to the var case". motion.g
mount-in fade preserved; only the AnimatePresence wrapper drops, since
there's no second branch to crossfade into. Container-query CSS classes
renamed in lockstep.
