+++ app/posts/the-function-that-remembered/widgets/index.ts
@@ -1,4 +1,5 @@
-export { LoopTrap } from "./LoopTrap";
-export { TetherScope } from "./TetherScope";
-export { RenderLoom } from "./RenderLoom";
+export { LoopLet } from "./LoopLet";
+export { LoopVar } from "./LoopVar";
 export { PredictReveal } from "./PredictReveal";
+export { RenderLoom } from "./RenderLoom";
+export { TetherScope } from "./TetherScope";
